[BUGFIX] Improve mount point index queue handling
When a page is stored in the Backend and a mount point page points
to it with overlay enabled the record monitor will now also add
the overlayed page to the index queue.

Pages with disabled mount point overlay will now be indexed as
normal pages by adjusting the additionalWhereClause config option.

Resolves: #59443
Change-Id: I3a0be14358161c7a6b5da5bed6cdde5e99714071